object_id ad u128

+++ benches/object_id.rs
@@ -6,5 +6,5 @@ use rbatis::object_id::ObjectId;
 
 #[bench]
 fn bench_object_id(b: &mut Bencher) {
-    b.iter(|| ObjectId::new().u64());
+    b.iter(|| ObjectId::new().u128());
 }
